lazy cast to float to stop warnings

+++ src/Song.cpp
@@ -442,7 +442,7 @@ bool Song::LoadFromSongDir(RString sDir, bool load_autosave, ProfileSlot from_pr
 LOG->Trace(
             "[LoadFromSongDir] File: %s, DefaultSyncOffset: %f, Group Offset: %f, "
             "Song Offset: %f",
-            GetSongFilePath().c_str(), PREFSMAN->m_DefaultSyncOffset,
+            GetSongFilePath().c_str(), (float)PREFSMAN->m_DefaultSyncOffset,
             SONGMAN->GetGroupFromName(m_sGroupName)->GetSyncOffset(),
             m_SongTiming.m_fBeat0GroupOffsetInSeconds);
 
@@ -679,7 +679,7 @@ void Song::TidyUpData( bool from_cache, bool /* duringCache */ )
         LOG->Trace(
             "[TidyUpData] File: %s, DefaultSyncOffset: %f, Group Offset: %f, "
             "Song Offset: %f",
-            GetSongFilePath().c_str(), PREFSMAN->m_DefaultSyncOffset,
+            GetSongFilePath().c_str(), (float)PREFSMAN->m_DefaultSyncOffset,
             SONGMAN->GetGroupFromName(m_sGroupName)->GetSyncOffset(),
             m_SongTiming.m_fBeat0GroupOffsetInSeconds);
